use the same logic for pypi API
